% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claim_info.R
\docType{data}
\name{bushel_claims}
\alias{bushel_claims}
\title{Phase Two - Bushel: Claims}
\format{
A tibble with 3,066 rows and 69 column variables:
}
\usage{
bushel_claims
}
\description{
The claim information for the research claims being assessed. This data is stored in wide format,
one row per claim, and includes additional metadata about the claim extraction process.
}
\examples{
bushel_claims
}
\keyword{datasets}
