% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bushel_priors.R
\docType{data}
\name{bushel_priors}
\alias{bushel_priors}
\title{Phase Two - Bushel: Priors}
\format{
A tibble with 3,066 rows and 2 column variables:
\describe{
\item{paper_id}{Unique, alphanumeric identifier for the paper/claim being assessed.}
\item{prior_means}{Logit-scale prediction of replicability.}
}
}
\usage{
bushel_priors
}
\description{
The prior means used in the BayPRIORsAgg aggregation method. This data is stored in long format with one
row per claim.
}
\examples{
bushel_priors
}
\keyword{datasets}
