% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claim_info.R
\docType{data}
\name{covid_claims}
\alias{covid_claims}
\title{Phase One - COVID-19 Single Trace: Claims}
\format{
A tibble with 100 rows and 79 column variables:
}
\usage{
covid_claims
}
\description{
The claim information for the research claims being assessed. This data is stored in wide format,
one row per claim, and includes additional metadata about the claim extraction process.
}
\examples{
covid_claims
}
\keyword{datasets}
