% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_cs.R
\docType{data}
\name{covid_cs}
\alias{covid_cs}
\title{Phase One - COVID-19 Single Trace: Confidence Scores}
\format{
A tibble with 3,800 rows and 4 column variables:
\describe{
\item{method}{Name of aggregation method that generated confidence score.}
\item{paper_id}{Unique, alphanumeric identifier for the paper/claim being assessed..}
\item{cs}{Aggregated confidence score.}
\item{n_experts}{Number of experts whose ratings were aggregated.}
}
}
\usage{
covid_cs
}
\description{
The confidence scores for each research claim question calculated via mathematical aggregation
of the participant ratings. This data is stored in long format with one row per aggregation
method and claim.
}
\examples{
covid_cs
}
\keyword{datasets}
