% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_priors.R
\docType{data}
\name{covid_priors}
\alias{covid_priors}
\title{Phase One - COVID-19 Single Trace: Priors}
\format{
A tibble with 100 rows and 2 column variables:
\describe{
\item{paper_id}{Unique, alphanumeric identifier for the paper/claim being assessed.}
\item{prior_means}{Logit-scale prediction of replicability.}
}
}
\usage{
covid_priors
}
\description{
The prior means used in the BayPRIORsAgg aggregation method. This data is stored in long format with one
row per claim.
}
\examples{
covid_priors
}
\keyword{datasets}
