% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claim_info.R
\docType{data}
\name{p1_claims}
\alias{p1_claims}
\title{Phase One - Single Trace: Claims}
\format{
A tibble with 3,000 rows and 78 column variables:
}
\usage{
p1_claims
}
\description{
The claim information for the research claims being assessed. This data is stored in wide format,
one row per claim, and includes additional metadata about the claim extraction process.
}
\examples{
p1_claims
}
\keyword{datasets}
