% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p1_priors.R
\docType{data}
\name{p1_priors}
\alias{p1_priors}
\title{Phase One - Single Trace: Priors}
\format{
A tibble with 3,000 rows and 2 column variables:
\describe{
\item{paper_id}{Unique, alphanumeric identifier for the paper/claim being assessed.}
\item{prior_means}{Logit-scale prediction of replicability.}
}
}
\usage{
p1_priors
}
\description{
The prior means used in the BayPRIORsAgg aggregation method. This data is stored in long format with one
row per claim.
}
\examples{
p1_priors
}
\keyword{datasets}
