% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p1_ratings.R
\docType{data}
\name{p1_ratings}
\alias{p1_ratings}
\title{Phase One - Single Trace: Ratings}
\format{
A tibble with 161,391 rows and 6 column variables:
\describe{
\item{round}{IDEA Protocol elicitation round. Either "round_1" or "round_2".}
\item{paper_id}{Unique, alphanumeric identifier for the paper/claim being assessed.}
\item{user_name}{Unique, anonymised, alphanumeric identifier for the participant.}
\item{question}{Question type.}
\item{element}{Component of response to question. Some questions use three-point elicitation.}
\item{value}{Participant's response to question. Ratings are reported on percentage scale (0-100), binary questions use -1/1 notation.}
}
}
\usage{
p1_ratings
}
\description{
The elicited numeric ratings of the participants. This data is stored in long format with one
row per participant, claim, elicitation round, and question.
}
\examples{
p1_ratings
}
\keyword{datasets}
