% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claim_info.R
\docType{data}
\name{p2_claims}
\alias{p2_claims}
\title{Phase Two - Single Trace: Claims}
\format{
A tibble with 9000 rows and 77 column variables:
}
\usage{
p2_claims
}
\description{
The claim information for the research claims being assessed. This data is stored in wide format,
one row per claim, and includes additional metadata about the claim extraction process.
}
\examples{
p2_claims
}
\keyword{datasets}
