% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_priors.R
\docType{data}
\name{p2_priors}
\alias{p2_priors}
\title{Phase Two - Single Trace: Priors}
\format{
A tibble with 900 rows and 2 column variables:
\describe{
\item{paper_id}{Unique, alphanumeric identifier for the paper/claim being assessed.}
\item{prior_means}{Logit-scale prediction of replicability.}
}
}
\usage{
p2_priors
}
\description{
The prior means used in the BayPRIORsAgg aggregation method. This data is stored in long format with one
row per claim.
}
\examples{
p2_priors
}
\keyword{datasets}
