% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_reasons.R
\docType{data}
\name{p2_reasons}
\alias{p2_reasons}
\title{Phase Two - Single Trace: Reasons}
\format{
A tibble with 4,164 rows and 19 column variables:
\describe{
\item{paper_id}{Unique, alphanumeric identifier for the paper/claim being assessed.}
\item{user_name}{Unique, anonymised, alphanumeric identifier for the participant.}
\item{RW:_Analysis_quality}{Number of times a participant used this reason on this paper.}
\item{RW:_Calibration_of_finding}{Number of times a participant used this reason on this paper.}
\item{RW:_Details_of_documentation}{Number of times a participant used this reason on this paper.}
\item{RW:_Domain_knowledge}{Number of times a participant used this reason on this paper.}
\item{RW:_Effect_size}{Number of times a participant used this reason on this paper.}
\item{RW:_Interaction_effect}{Number of times a participant used this reason on this paper.}
\item{RW:_Measurement_reliability}{Number of times a participant used this reason on this paper.}
\item{RW:_Plausibility}{Number of times a participant used this reason on this paper.}
\item{RW:_Population_or_subject_characteristics}{Number of times a participant used this reason on this paper.}
\item{RW:_Publication_date}{Number of times a participant used this reason on this paper.}
\item{RW:_Reporting_of_power_or_sample_size_calculations}{Number of times a participant used this reason on this paper.}
\item{RW:_Reputation_of_an_author,_team,_or_institution}{Number of times a participant used this reason on this paper.}
\item{RW:_Reputation_of_field_of_research}{Number of times a participant used this reason on this paper.}
\item{RW:_Reputation_of_journal}{Number of times a participant used this reason on this paper.}
\item{RW:_Sample_size_or_power_adequacy}{Number of times a participant used this reason on this paper.}
\item{RW:\emph{Statistical_significance}(p-values,_etc)}{Number of times a participant used this reason on this paper.}
\item{RW:_Uncertainty,_statistical_measures}{Number of times a participant used this reason on this paper.}
}
}
\usage{
p2_reasons
}
\description{
Participant justifications were analysed against a code-book to identify the distinct reasons
that informed their ratings. These reasons were assessed by multiple coders and inter-coder
reliability (ICR) was evaluated and only the reason codes passing these tests were retained.
This data is stored in a wide format table with one row per participant and claim, and one
column per reason code.
}
\examples{
p2_reasons
}
\keyword{datasets}
