% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repliCATS-package.R
\docType{package}
\name{repliCATS-package}
\alias{repliCATS}
\alias{repliCATS-package}
\title{repliCATS: Access The repliCATS-SCORE datasets}
\description{
This is a data package to access all publicly released repliCATS datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://https://github.com/metamelb-repliCATS/repliCATS}
  \item Report bugs at \url{https://github.com/metamelb-repliCATS/repliCATS/issues}
}

}
\author{
\strong{Maintainer}: David Wilkinson \email{david.wilkinson.research@gmail.com} (\href{https://orcid.org/0000-0002-9560-6499}{ORCID})

}
\keyword{internal}
