% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPS_priors.R
\docType{data}
\name{SIPS_priors}
\alias{SIPS_priors}
\title{SIPS - Known Outcomes: Priors}
\format{
A tibble with 25 rows and 2 column variables:
\describe{
\item{paper_id}{Unique, alphanumeric identifier for the paper/claim being assessed.}
\item{prior_means}{Logit-scale prediction of replicability.}
}
}
\usage{
SIPS_priors
}
\description{
The prior means used in the BayPRIORsAgg aggregation method. This data is stored in long format with one
row per claim.
}
\examples{
SIPS_priors
}
\keyword{datasets}
